Reduce CPU usage in viz pipeline bgo#555834
Overhaul the visualization pipeline again. Allow disabling of the pipeline to
conserve CPU when the data is not being used. Discard the spectrum element and
use libgstfft directly to eliminate complex timing and synchronization issues.

The pipeline is now enabled by default, but no processing is done until a
handler is connected on the managed PlayerEngine.

Build is updated to link against libgstfft.