[TaskStatusIcon] API/functionality improvements
Support being visible when any kind of task is running, not just
background ones. Also support the ability to enable or disable
the intermittent visibility of the spinner, and configure its
timeouts. This is useful for clients (e.g. Cubano) that do not
display the actual UserJobTiles in their main UI.

Fixes also potential race issues with the intermittent timeouts.

Clean up the tooltip a bit, and udpate it when individual jobs
update (e.g. reflect job title changes in the tooltip).