[PlayerEngine] Added Initialize/DelayedInitialize
PlayerEngine now has an Initialize method which can be invoked
delayed by PlayerEngineService.DelayedInitialize
(ala IDelayedInitializeService). Just before a PlayerEngine is
initialized, the EngineBeforeInitialize event is raised on the
PlayerEngineService, and after it is initialized, the
EngineAfterInitialize event is raised.

This provides hooks for other services to configure the engine
before it is constructed, and interact with it after it is
safely constructed.