Reorder audiotee pad allocation (fixes BGO #583384)
In GStreamer 0.10.21-0.10.22, the last source pad allocated on a tee is used
for buffer allocations.  When the vis pipeline is disabled, this caused a
freeze since it was attached to the last-allocated pad.

This fix simply ensures that the pad linked to the hardware sink is the last
one allocated.

Signed-off-by: Aaron Bockover <abockover@novell.com>