[GStreamer PlayerEngine] Work around pipeline bug
There's still a regression somewhere in the engine/pipeline that is resulting
in an invalid first pipeline if the first track for the engine is video.

The pipeline destruction in case of this error caused the clutter video
sink to be reset. To avoid this issue for now, we always create a new video
sink when the pipeline is constructed.

This needs closer investigation.