[ScreensaverManager] Moved from Banshee.NowPlaying to Banshee.Services
This code is not specific to the Now Playing extensions, is useful
elsewhere, especially when the new Now Playing extension that will
be merged from Cubano is on its way.

The real problem was the hard dependency the GNOME backend had on
the legacy Now Playing extension. It made it not possible to remove
the legacy Now Playing from a build while still retaining GNOME
features, like GConf support.