[PixbufImageSurface] Add convenience factory func
There are many places were we create PixbufImageSurface (all touched in
this patch) where we assume that nothing is thrown, but that null is
returned if there is an error.  This patch adds a convenience factory
method (only adding to the API; all ctors are left alone) that makes
sure the passed in Pixbuf's Handle isn't null, and will catch any errors
in creating the PixbufImageSurface and return null.