[GtkElementService] Add content window APIs
Three new APIs that allow extensions/clients to register arbitrary
content windows. A content window is a window that immerses the
user in media content (e.g. the Moblin media panel).

Provides RegisterContentWindow, UnregisterContentWindow, and
an enumerable ContentWindows property.

Patch the NotificationAreaService to iterate over ContentWindows
instead of just checking the PrimaryWindow for toplevel focus
to determine if it is appropriate to show a notification.