[QueueableSourceComboBox] Use CellRendererText
The SourceRowRenderer cell renderer has a bug where
width reporting does not work correctly, so it's fixed
to 200px wide. This causes the QueueableSourceComboBox
to also be 200+px wide, regardless of the text width
of the sources in the box, causing problems with UI
like Muinshee and the Moblin Media Panel.

Revert to just using a stock CellRendererText for now.