[Nereid] Re-work some of the UI for Moblin
This doesn't actually change the way Nereid looks, but
it does some packing differently, so that it can be
manipulated more easily by the Moblin extension.

It also removes the footer toolbar in favor of just a
normal HBox since there's no real code utilizing the
extension slot, and it's a theming nightmare.