Migrate to the Last.fm 2.0 API (BGO#541227)
This commit upgrades the Lastfm to use the 2.0 API, as the older API was
disabled a while back. Radio streams are accessible again, under
restrictions imposed by Last.fm : paying subscribers and people
living the US, UK or Germany.

Major changes are :
- Web services style API returning JSON, handled in the new
LastfmRequest class.
- New authentication scheme. The user's doesn't have to give us
his password, but has to authorize Banshee to access his account.
- Scrobbling uses this authentication scheme, moving it to the 1.2.1
protocol.
The Lastfm.Data namespace is not impacted.