[ListView] Fix column size calculation
RecalculateColumnSizes() stops adapting the column sizes when
first value of remaining_width is negative. This always happens if
there are several fixed-width columns (e.g. year, bitrate and rating)
at the end of the list view. The result is that these columns are
allocated less space than they require.

The fix changes the method so that it continues refitting the columns
until remaining_width is close to zero.